+++ test/test_kafka-consumer.rb
@@ -16,7 +16,7 @@ class TestKafkaConsumer < Test::Unit::TestCase
     send_test_messages topics[0]
     queue = SizedQueue.new(20)
     consumer = Kafka::KafkaConsumer.new(options)
-    consumer.subscribe
+    consumer.subscribe topics
     runner_thread = Thread.new { kafka_consumer_test_blk consumer, queue}
     begin
       timeout(30) do